refactor(investiment):move the archives to investiment package
