+++ backend/src/main/java/com/brokendev/backend/account/service/AccountService.java
@@ -11,8 +11,8 @@ import com.brokendev.backend.account.dto.TransactionStatementResponseDTO;
 import com.brokendev.backend.common.exceptions.InvalidDepositAmountException;
 import com.brokendev.backend.enums.TransactionType;
 import com.brokendev.backend.common.exceptions.UserAccountNotFoundException;
+import com.brokendev.backend.investment.domain.InvestmentRepository;
 import com.brokendev.backend.pix_transfer.domain.PixTransactionRepository;
-import com.brokendev.backend.repositories.*;
 import com.brokendev.backend.services.NotificationService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
