+++ backend/src/main/java/com/brokendev/backend/investment/domain/Investment.java
@@ -1,4 +1,4 @@
-package com.brokendev.backend.domain;
+package com.brokendev.backend.investment.domain;
 
 import com.brokendev.backend.account.domain.Account;
 import com.brokendev.backend.enums.InvestmentType;
