+++ backend/src/main/java/com/brokendev/backend/investment/dto/InvestmentResponseDTO.java
@@ -1,4 +1,4 @@
-package com.brokendev.backend.dto.investment;
+package com.brokendev.backend.investment.dto;
 
 import com.brokendev.backend.enums.InvestmentType;
