+++ backend/src/test/java/com/brokendev/backend/account/service/AccountServiceTest.java
@@ -3,7 +3,6 @@ package com.brokendev.backend.account.service;
 import com.brokendev.backend.account.domain.Account;
 import com.brokendev.backend.account.domain.AccountRepository;
 import com.brokendev.backend.account.dto.AccountBalanceResponseDTO;
-import com.brokendev.backend.account.dto.AccountDepositRequestDTO;
 import com.brokendev.backend.account.dto.AccountDepositResponseDTO;
 import com.brokendev.backend.account.dto.TransactionStatementResponseDTO;
 import com.brokendev.backend.boleto_payment.domain.BoletoPayment;
@@ -12,11 +11,11 @@ import com.brokendev.backend.common.domain.user.User;
 import com.brokendev.backend.common.domain.user.UserRepository;
 import com.brokendev.backend.common.exceptions.InvalidDepositAmountException;
 import com.brokendev.backend.common.exceptions.UserAccountNotFoundException;
-import com.brokendev.backend.domain.Investment;
+import com.brokendev.backend.investment.domain.Investment;
 import com.brokendev.backend.enums.*;
 import com.brokendev.backend.pix_transfer.domain.PixTransaction;
 import com.brokendev.backend.pix_transfer.domain.PixTransactionRepository;
-import com.brokendev.backend.repositories.InvestmentRepository;
+import com.brokendev.backend.investment.domain.InvestmentRepository;
 import com.brokendev.backend.services.NotificationService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -27,7 +26,6 @@ import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 
-import javax.security.auth.login.AccountNotFoundException;
 import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.Arrays;
